Fix/frontend apis

+++ backend/src/main/java/com/org/backend/service/DepartmentService.java
@@ -8,6 +8,7 @@ import com.org.backend.entity.Department;
 import com.org.backend.entity.Employee;
 import com.org.backend.repository.DepartmentRepository;
 import com.org.backend.repository.EmployeeRepository;
+import com.org.backend.repository.PositionRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,6 +22,8 @@ public class DepartmentService {
 
     private final DepartmentRepository departmentRepository;
     private final EmployeeRepository employeeRepository;
+    private final EmployeeService employeeService;
+    private final PositionRepository positionRepository;
 
     public List<DepartmentDto> getAllDepartments(){
 
@@ -44,7 +47,7 @@ public class DepartmentService {
 
         Department department = new Department();
         department.setName(request.name());
-        department.setDepartmentCode(request.departmentCode());
+        department.setDepartmentCode(request.code());
 
         if (request.parentDepartmentId() != null){
             Department parentDepartment = departmentRepository
@@ -77,7 +80,7 @@ public class DepartmentService {
                 .orElseThrow(() -> new IllegalArgumentException("Invalid department id"));
 
         department.setName(request.name());
-        department.setDepartmentCode(request.departmentCode());
+        department.setDepartmentCode(request.code());
 
         if (request.parentDepartmentId() != null){
             Department parentDepartment = departmentRepository
@@ -104,6 +107,10 @@ public class DepartmentService {
 
         Department department = departmentRepository.findById(departmentId)
                 .orElseThrow(() -> new IllegalArgumentException("Invalid department id"));
+
+        if (positionRepository.countByDepartment(department) > 0){
+            throw new IllegalArgumentException("Positions are mapped to this department !");
+        }
 
         // soft delete logic in future
         departmentRepository.delete(department);
@@ -133,7 +140,8 @@ public class DepartmentService {
                 department.getId(),
                 department.getName(),
                 department.getDepartmentCode(),
-                department.getParentDepartment() != null ? department.getParentDepartment().getId() : null
+                department.getParentDepartment() != null ? department.getParentDepartment().getId() : null,
+                department.getHeadEmployee() != null ? employeeService.mapToDto(department.getHeadEmployee()) : null
         );
     }
 }
